+++ tests/search/tensor_eval/unstable-docs.json
@@ -1,5 +1,5 @@
 [
-  { "put": "id:multi:multi::1", "fields": {
+  { "put": "id:unstable:unstable::1", "fields": {
       "id": 1,
       "title": "barbie",
       "mydoubles" : { "cells": { "barbie": 5.25 } },
@@ -8,7 +8,7 @@
       "mybytes" : { "cells": { "barbie": 2.0 } }
     }
   },
-  { "put": "id:multi:multi::2", "fields": {
+  { "put": "id:unstable:unstable::2", "fields": {
       "id": 2,
       "title": "heman",
       "mydoubles" : { "cells": { "heman": 7.25 } },
@@ -17,7 +17,7 @@
       "mybytes" : { "cells": { "heman": 3.0 } }
     }
   },
-  { "put": "id:multi:multi::3", "fields": {
+  { "put": "id:unstable:unstable::3", "fields": {
       "id": 3,
       "title": "tv",
       "mydoubles" : { "cells": { "tv": 8.25 } },
