+++ tests/search/tensor_eval/unstable_search/query-profiles/default.xml
@@ -1,0 +1,1 @@
+<query-profile id="default" type="root" />
